purchase 오류수정완료
gamedetail에서 구매하기 장바구니 넘어가는 오류 수정, 장바구니에서 한개만 넘겼을때 오류 수정, 한개만 구매했을때 데이터 들어가는 오류 수정완료

+++ playfarmb/src/main/java/com/example/playfarmb/store/service/PurchaseServiceImpl.java
@@ -81,17 +81,19 @@ public class PurchaseServiceImpl implements PurchaseService {
 	    Purchaselist savedPurchase = purchaseRepository.save(purchase);
 
 	    // 각 게임에 대한 Listdetail 생성 및 추가
-	    for (Listdetail listDetailDTO : dto.getListDetails()) {
-	        Listdetail listDetail = Listdetail.builder()
-	            .purchId(new ListdetailId(savedPurchase.getPurchId(), listDetailDTO.getPurchId().getGameId(), listDetailDTO.getPurchId().getPlaytype()))
-	            .purchaselist(savedPurchase)
-	            .build();
-
-	        savedPurchase.getListDetails().add(listDetail);
-	        
-	        Cart cartlist = cartRepository.findById(new CartId(dto.getUserId(), listDetailDTO.getPurchId().getGameId(), listDetailDTO.getPurchId().getPlaytype()))
-	        		.orElseThrow(() -> new RuntimeException("카트에 같은 게임 정보가 없습니다"));
-	        cartRepository.delete(cartlist);
+	    if (dto.getListDetails() != null && !dto.getListDetails().isEmpty()) {
+	    	for (Listdetail listDetailDTO : dto.getListDetails()) {
+	    		Listdetail listDetail = Listdetail.builder()
+	    				.purchId(new ListdetailId(savedPurchase.getPurchId(), listDetailDTO.getPurchId().getGameId(), listDetailDTO.getPurchId().getPlaytype()))
+	    				.purchaselist(savedPurchase)
+	    				.build();
+	    		
+	    		savedPurchase.getListDetails().add(listDetail);
+	    		
+	    		Cart cartlist = cartRepository.findById(new CartId(dto.getUserId(), listDetailDTO.getPurchId().getGameId(), listDetailDTO.getPurchId().getPlaytype()))
+	    				.orElseThrow(() -> new RuntimeException("카트에 같은 게임 정보가 없습니다"));
+	    		cartRepository.delete(cartlist);
+	    	}
 	    }
 	    
 	    // Listdetail 추가 후 다시 저장
